planクラスを別ファイルへ分離
