+++ challengeA/okamoto/spec/plan_spec.rb
